A Concourse CI Resource for VMware NSX-T version
This Concourse resource has but one purpose: to extract the version
number of the NSX-T infrastructure when passed the URL of the NSX-T
manager along with appropriate credentials.

This can be used in pipelines when it's important to record the version
number of the NSX-T infrastructure (for compatibility purposes, say).

[#162397653](https://www.pivotaltracker.com/story/show/162397653)

Signed-off-by: Brian Cunnie <bcunnie@pivotal.io>